requirements.txt
fastapi
uvicorn
google-generative-ai
pydantic